add mock email sender
